Move the BedrockRawChunk class
