+++ src/amulet/level/bedrock/chunk.pyi
@@ -1,17 +1,8 @@
 from __future__ import annotations
 
-import collections.abc
-
 import amulet.core.chunk
-import amulet.nbt
 
-__all__: list[str] = ["BedrockChunk", "BedrockRawChunk"]
+__all__: list[str] = ["BedrockChunk"]
 
 class BedrockChunk(amulet.core.chunk.Chunk):
     pass
-
-class BedrockRawChunk:
-    @property
-    def actors(self) -> collections.abc.MutableSequence[amulet.nbt.NamedTag]: ...
-    @property
-    def data(self) -> collections.abc.MutableMapping[bytes, bytes]: ...
